requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
google-auth==2.28.0
google-cloud-aiplatform==1.42.1
h11==0.14.0
idna==3.6
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pytesseract==0.3.10
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
